chore: sunset internal, device logic on main thread
