+++ apps/ledger-live-desktop/package.json
@@ -123,7 +123,6 @@
     "styled-system": "^5.1.5",
     "timemachine": "^0.3.2",
     "tippy.js": "^6.3.7",
-    "tree-kill": "^1.2.2",
     "uncontrollable": "^7.2.1",
     "uuid": "^8.3.2",
     "winston": "^3.2.1",
