+++ apps/ledger-live-desktop/src/index.js
@@ -7,10 +7,4 @@ if (getEnv("PLAYWRIGHT_RUN")) {
   });
 }
 
-if (!process.env.IS_INTERNAL_PROCESS) {
-  // Main electron thread
-  require("./main");
-} else {
-  // Internal thread (coins, hardware)
-  require("./internal");
-}
+require("./main");
