+++ apps/ledger-live-desktop/src/main/device/live-common-setup.js
