+++ apps/ledger-live-desktop/src/sentry/pname.js
@@ -2,11 +2,6 @@
 
 // Infer a "pname" aka short id version of process name
 
-const pname =
-  typeof window === "undefined"
-    ? process.env.IS_INTERNAL_PROCESS
-      ? "internal"
-      : "main"
-    : "renderer";
+const pname = typeof window === "undefined" ? "main" : "renderer";
 
 export default pname;
