+++ package.json
@@ -39,7 +39,7 @@
     "watch:common": "pnpm turbo watch --no-daemon --filter=live-common",
     "dev:cli": "pnpm turbo watch --no-daemon --filter=live-cli",
     "dev:lld": "pnpm turbo start --no-daemon --filter=ledger-live-desktop",
-    "dev:lld:debug": "DEV_TOOLS=1 LEDGER_INTERNAL_ARGS=--inspect ELECTRON_ARGS=--remote-debugging-port=8315 pnpm turbo start --no-daemon --filter=ledger-live-desktop",
+    "dev:lld:debug": "DEV_TOOLS=1 ELECTRON_ARGS=--remote-debugging-port=8315 pnpm turbo start --no-daemon --filter=ledger-live-desktop",
     "dev:llm": "pnpm turbo start --no-daemon --filter=live-mobile",
     "release:lld": "pnpm turbo release --no-daemon --filter=ledger-live-desktop",
     "pre:lld": "pnpm turbo pre-build --no-daemon --filter=ledger-live-desktop",
